% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AT_Stress_Space}
\alias{AT_Stress_Space}
\title{Arabidopsis StressSpace Based on ATH1 Chip GEO Data}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=[GSEXXXX]} Replace GSEXXXX by one of the
actual dataset series names listed in 'Details' section.
}
\usage{
data(AT_Stress_Space)
}
\description{
AT_Stress_Space spans a 14-dimension physiological space constructed from
50 different Arabidopsis thaliana stress-response gene expression datasets
from "Gene Expression Omnibus" (GEO). All datasets are measured by Affymetrix
Arabidopsis ATH1 Genome Array.
}
\details{
Here's the list of datasets used for making this Space: GSE5539, GSE5615,
GSE5616, GSE5620-5628, GSE5727, GSE6177, GSE6583, GSE7334, GSE7642,
GSE8787, GSE10643, GSE10670, GSE12029, GSE12619, GSE15577, GSE16474,
GSE16765, GSE18217, GSE18666, GSE19700, GSE22107, GSE22671, GSE24177,
GSE26983, GSE26983, GSE27861, GSE35258, GSE40061, GSE41626, GSE44655,
GSE46107, GSE39236, GSE46205, GSE40998&43724, GSE40998&43724, GSE44922,
GSE44053, GSE55431, GSE58616, GSE58620 and GSE39384.

The difference between AT_Stress_Space and AT_Stress_Space_detailed is that
in AT_Stress_Space all stressed dimensions with the same stress type are
averaged to make a universal axis for each stress type, while in
the space of AT_Stress_Space_detailed, stress responses from different
datasets havetheir own dedicated dimension.
}
\keyword{datasets}
