% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AT_Stress_Space_Meta}
\alias{AT_Stress_Space_Meta}
\title{Arabidopsis thaliana Meta-Stress Space Based on ATH1 Chip GEO Data}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\usage{
data(AT_Stress_Space_Meta)
}
\description{
Almost identical to AT_Stress_Space, except in this space in place of
normal stress groups, meta stress groups are used. Meta groups are found in the
paper 'under preparation' (email esfahani(at)aices.rwth-aachen.de to get
access to a draft/link).
}
\keyword{datasets}
