% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AT_Stress_Space_RNASeq}
\alias{AT_Stress_Space_RNASeq}
\title{Arabidopsis thaliana Stress Space Based on RNA-seq SRA Data}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/[PRJXXXX]} (Replace PRJXXXX by one of the
actual dataset series names listed in 'Details' section.)
}
\usage{
data(AT_Stress_Space_RNASeq)
}
\description{
AT_Stress_Space_RNASeq spans a 7-dimension physiological space
constructed from 8 different Arabidopsis thaliana stress-response
gene expression datasets from "Sequence Read Archive" (SRA).
}
\details{
Here's the list of datasets used for making this Space:
PRJEB13938 (E.MTAB.4683), PRJNA232536 (GSE53641), PRJNA263595 (GSE62256),
PRJNA272425 (GSE64870), PRJNA315516, PRJNA319318 (GSE80565),
PRJNA348676 (GSE88798), PRJNA412447

The difference between AT_Stress_Space_RNASeq and
AT_Stress_Space_RNASeq_detailed is that in AT_Stress_Space_RNASeq all
stressed dimensions with the same stress type are
summarized to make a universal axis for each stress type, while in
the space of AT_Stress_Space_RNASeq_detailed stress responses from different
datasets have their own dedicated dimension.
}
\keyword{datasets}
