% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GM_Stress_Space}
\alias{GM_Stress_Space}
\title{Glycine max Stress Space Based on Affymetrix Soybean Genome Array GEO Data}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=[GSEXXXX]} (Replace GSEXXXX by one of the
actual dataset series names listed in 'Details' section.)
}
\usage{
data(GM_Stress_Space)
}
\description{
GM_Stress_Space spans a 6-dimension physiological space
constructed from 11 different Soybean stress-response
gene expression datasets from "Gene Expression Omnibus" (GEO).
}
\details{
Here's the list of datasets used for making this Space:
GSE18423, GSE18517, GSE18518, GSE22158, GSE29663, GSE32642, GSE40604,
GSE50408, GSE51328, GSE51710, GSE55896

The difference between GM_Stress_Space and
GM_Stress_Space_Detailed is that in GM_Stress_Space all
stressed dimensions with the same stress type are
summarized to make a universal axis for each stress type, while in
the space of GM_Stress_Space_Detailed stress responses from different
datasets have their own dedicated dimension.
}
\keyword{datasets}
