% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OS_Stress_Space}
\alias{OS_Stress_Space}
\title{Oryza sativa Stress Space Based on Affymetrix Rice Genome Array GEO Data}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=[GSEXXXX]} (Replace GSEXXXX by one of the
actual dataset series names listed in 'Details' section.)
}
\usage{
data(OS_Stress_Space)
}
\description{
OS_Stress_Space spans a 4-dimension physiological space
constructed from 8 different Rice stress-response
gene expression datasets from "Gene Expression Omnibus" (GEO).
}
\details{
Here's the list of datasets used for making this Space:
GSE14403, GSE15448, GSE16108, GSE21651, GSE24048, GSE25176, GSE26280,
GSE6901

The difference between OS_Stress_Space and
OS_Stress_Space_Detailed is that in OS_Stress_Space all
stressed dimensions with the same stress type are
summarized to make a universal axis for each stress type, while in
the space of OS_Stress_Space_Detailed stress responses from different
datasets have their own dedicated dimension.
}
\keyword{datasets}
