% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TA_Stress_Space}
\alias{TA_Stress_Space}
\title{Triticum aestivum Stress Space Based on Affymetrix Wheat Genome Array GEO Data}
\format{A matrix with genes as rows and different physiological dimensions in columns.
Gene Entrez IDs are stored in rownames and space-dimension names in colnames.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=[GSEXXXX]} (Replace GSEXXXX by one of the
actual dataset series names listed in 'Details' section.)
}
\usage{
data(TA_Stress_Space)
}
\description{
TA_Stress_Space spans a 2-dimension physiological space
constructed from 4 different Wheat stress-response
gene expression datasets from "Gene Expression Omnibus" (GEO).
}
\details{
Here's the list of datasets used for making this Space:
GSE11774, GSE14697, GSE30436, GSE31759

The difference between TA_Stress_Space and
TA_Stress_Space_Detailed is that in TA_Stress_Space all
stressed dimensions with the same stress type are
summarized to make a universal axis for each stress type, while in
the space of TA_Stress_Space_Detailed stress responses from different
datasets have their own dedicated dimension.
}
\keyword{datasets}
